feat!: introduce a new data model and software components based on PURL (#305)
This PR changes the data model in Macaron to analyze at software component level using [PURL](https://github.com/package-url/purl-spec) as the identifier instead of repository level analysis. In this new data model, it is allowed to miss a repository associated with a component, closing this [issue](https://github.com/oracle/macaron/issues/187).

In addition, it closes this [issue](https://github.com/oracle/macaron/issues/115) by changing the database transactions to guarantee atomicity per analysis, i.e., if anything goes wrong and the analysis fails unexpectedly, no changes will be added to the database.

In this PR the input is still a repository, and a PURL is constructed from the repository provided as input. In a subsequent PR, we will add [support for PURL identifier to be provided as a CLI input](https://github.com/oracle/macaron/issues/375).

Signed-off-by: behnazh-w <behnaz.hassanshahi@oracle.com>