chore: add changesets/action for npm release (#503)
This PR adds `changesets/action` to the list of GitHub Actions used to publish
npm packages. An example project that uses this Action to automate releases 
is `sigstore/sigstore-js`. Note that there is a TODO to also check the `publish` input 
provided to this Action to improve detection accuracy.

This PR also fixes a bug for using the deploy `github_actions` for build tools.
Previously,  only the `github_actions` of pip build tool was read from 
`defaults.ini` and used in  `build_as_code_check.py`. This fix removes this 
hard-coded read from `defaults.ini` and makes it applicable to all build tools.

Signed-off-by: behnazh-w <behnaz.hassanshahi@oracle.com>