refactor!: allow the branch name in the schema of a repository to be null (#532)
This PR changes the Repository schema in the data model to allow a branch name to be None. This is in preparation for mapping artifacts to commits because sometimes we might not be able to find a branch. However, we should still be able to checkout using the resolved commit hash.

Note that the local `macaron.db` needs to be removed to apply this change in the data model, hence the breaking change scope in the PR title.

Signed-off-by: behnazh-w <behnaz.hassanshahi@oracle.com>