test: fix the GitHub Actions workflow detection test (#591)
This PR fixes a bug in the GitHub Actions workflow detection test. The assertions in the test_get_workflows function were checking the return values of list.sort(), which is an in-place function and doesn't return anything. Instead we should compare the sorted lists.

This PR also refactors the tests and removes MacaronTestCase, which is not necessary anymore.

Signed-off-by: behnazh-w <behnaz.hassanshahi@oracle.com>