feat: use provenance to find commits for supported PURL types. (#653)
This PR allows for the extraction of repository URLs and related commits from provenance files. Supported provenance includes SLSA v0.1, 0.2, and 1, as well as Witness v0.1. This feature takes effect when a user supplies a provenance as input to the analysis, or when one can be retrieved from npm or a configured JFrog repository, as applicable.

Signed-off-by: Ben Selwyn-Smith <benselwynsmith@googlemail.com>