feat: allow provenance files to be files containing a URL pointing to the actual provenance file which will be transparently downloaded
URL link file format is the Windows .url file format i.e. an ini file containing a "URL" field inside an "InternetShortcut" section.
When processing the URL link provenance file, the content will be transparently downloaded if the hostname of the URL is configured as an allowed hostname in the defaults.ini file.

Signed-off-by: Nicholas Allen <nicholas.allen@oracle.com>