fix: do not fetch from origin/HEAD for local repo targets (#734)
For analyzing local repositories in offline mode, we should not fetch from origin/HEAD because user might want to run Macaron on the local changes. This PR changes the check_out_repo_target function to fix this issue: if branch_name and a commit are not provided, this function will not do anything and the HEAD commit will be analyzed. If there are uncommitted local changes, the HEAD commit will appear in the report but the repo with local changes will be analyzed. We leave it up to the user to decide whether to commit the changes or not.

Signed-off-by: behnazh-w <behnaz.hassanshahi@oracle.com>