feat: verify npm SLSA provenance against signed npm provenance (#747)
This PR alters the behaviour of interactions with npm registries. When possible, the "publish" attestation will be downloaded alongside the usual SLSA provenance, allowing the unsigned SLSA provenance to be verified by matching it against the signed "publish" attestation. A new check has been added to report on this new feature.

Signed-off-by: Ben Selwyn-Smith <benselwynsmith@googlemail.com>