feat: add command to run repo and commit finder without analysis (#827)
This PR adds a new command, find-source, that runs the commit finder and repo finder on a given PURL whilst skipping analysis. It also optionally accepts a repo path, in which case only the commit finder will be called. The results of the operation show up in the logs, and are written to a file in a JSON report format.

Signed-off-by: Ben Selwyn-Smith <benselwynsmith@googlemail.com>