feat(heuristics): add Fake Email analyzer to validate maintainer email domain (#1106)
This PR adds a new heuristic analyzer called FakeEmailAnalyzer. It verifies the validity of maintainer email addresses listed in a PyPI package by checking both the format and the existence of records for their domains. This helps detect packages with fake or throwaway emails, which are often indicative of malicious intent.

Signed-off-by: Amine <amine.raouane@enim.ac.ma>