docs: update provenance tutorial (#1110)
This PR updates the npm provenance discovery tutorial to include the newer discovery methods: PyPI and GitHub. The tutorial is updated to provide examples of provenance discovery for PyPI and GitHub using the toga PyPI library, which has provenance of either type available for discovery depending on the exact version specified. For GitHub attestation, a small code update is included to allow the repository URL and commit digest to be found from provenance. Also, the verify-provenance command line argument has been removed, making this action occur by default once again.

Signed-off-by: Ben Selwyn-Smith <benselwynsmith@googlemail.com>