fix: improve build tool detection (#1169)
This PR refactores the build tool detection logic by reusing the repo verifier’s path filtering approach. This change ensures that paths that are likely used for testing are excluded during the build tool detection process. Additionally, this PR adds an integration test for a project that uses Maven, with Gradle used as an example to demonstrate the filtering mechanism.

Signed-off-by: behnazh-w <behnaz.hassanshahi@oracle.com>