chore: improving efficiency of similar projects analyzer (#1170)
The goal is to improve the efficiency of the SimilarProjectsAnalyzer, which currently downloads the sourcecode tarball for every package of every maintainer it finds. The solution to this is to use the structure of the tarball/wheel provided on the inspector.pypi.io page of the package, making web requests to extract the structure instead of downloading the package.

Signed-off-by: Carl Flottmann <carl.flottmann@oracle.com>